lab2/homework/requirements.txt
aiodns==3.0.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
bcrypt==4.0.1
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==39.0.2
ecdsa==0.18.0
fastapi==0.94.1
frozenlist==1.3.3
h11==0.14.0
hiredis==2.2.2
httptools==0.5.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
np==1.0.2
numpy==1.24.2
passlib==1.7.4
Pillow==9.4.0
pyasn1==0.4.8
pycares==4.3.0
pycparser==2.21
pydantic==1.10.6
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0
redis==4.5.1
requests==2.28.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
yarl==1.8.2
